requirements.txt
kafka-python==2.3.0
py4j==0.10.9.9
pyspark==4.1.1
